refactor(sdk/reader): move codex parse engine into codex/incremental.rs
Mirror the claude split: extract the codex incremental-parse engine —
Pending<T>, CodexParseState + its handlers, CommittedSnapshot, parse_codex_buffer
— out of the oversized codex.rs (1801) into codex/incremental.rs, leaving the
root as the public API + parse-primitive helpers. codex.rs drops to 831;
incremental.rs is 1014.

Pure code-movement: the moved block is byte-identical (only parse_codex_buffer
gained pub(super)); the engine's root dependencies were widened to
pub(in crate::reader::codex) and now-unused root imports pruned. All three
readers now share one shape — each is a directory with its engine in
incremental.rs. 47 codex tests pass, full SDK green, clippy clean, live OK.
