+++ crates/relayburn-sdk/src/reader/codex.rs
@@ -10,27 +10,22 @@
 //! sizing. The Rust port uses [`HeuristicCounter`] (bytes/4) — see #246 for
 //! the cl100k hookup.
 
-use std::collections::{BTreeMap, BTreeSet, HashMap};
+use std::collections::{BTreeSet, HashMap};
 use std::fs::File;
-use std::io::{BufRead, BufReader, Read, Seek, SeekFrom};
+use std::io::{BufReader, Read, Seek, SeekFrom};
 use std::path::Path;
 
 use serde_json::Value;
 
-use crate::reader::classifier::{classify_activity, ClassificationInput};
 use crate::reader::fidelity::classify_fidelity;
 use crate::reader::git::ProjectResolver;
-use crate::reader::hash::{args_hash, content_hash};
+use crate::reader::hash::content_hash;
 use crate::reader::types::{
-    CompactionEvent, ContentKind, ContentRecord, ContentRole, ContentStoreMode, ContentToolResult,
-    ContentToolUse, Coverage, Fidelity, RelationshipSourceKind, RelationshipType,
-    SessionRelationshipRecord, SourceKind, ToolCall, ToolResultEventRecord, ToolResultEventSource,
-    ToolResultStatus, TurnRecord, Usage, UsageGranularity, UserTurnBlock, UserTurnBlockKind,
-    UserTurnRecord,
+    CompactionEvent, ContentRecord, ContentStoreMode, Coverage, Fidelity, RelationshipSourceKind,
+    RelationshipType, SessionRelationshipRecord, SourceKind, ToolCall, ToolResultEventRecord,
+    ToolResultStatus, TurnRecord, Usage, UsageGranularity, UserTurnBlock, UserTurnRecord,
 };
-use crate::reader::user_turn::{
-    join_nonempty, resolve_token_counter, HeuristicCounter, UserTurnTokenizer,
-};
+use crate::reader::user_turn::{resolve_token_counter, UserTurnTokenizer};
 
 // ---------------------------------------------------------------------------
 // Public surface
@@ -215,21 +210,21 @@ pub fn parse_codex_session_incremental(
 // ---------------------------------------------------------------------------
 
 #[derive(Debug, Clone, Default)]
-struct UserTurnSlot {
-    blocks: Vec<UserTurnBlock>,
-    preceding_message_id: Option<String>,
-    ts: String,
+pub(in crate::reader::codex) struct UserTurnSlot {
+    pub(in crate::reader::codex) blocks: Vec<UserTurnBlock>,
+    pub(in crate::reader::codex) preceding_message_id: Option<String>,
+    pub(in crate::reader::codex) ts: String,
 }
 
 impl UserTurnSlot {
-    fn from_persisted(p: &PersistedUserTurnSlot) -> Self {
+    pub(in crate::reader::codex) fn from_persisted(p: &PersistedUserTurnSlot) -> Self {
         Self {
             blocks: p.blocks.clone(),
             preceding_message_id: p.preceding_message_id.clone(),
             ts: p.ts.clone(),
         }
     }
-    fn to_persisted(&self) -> PersistedUserTurnSlot {
+    pub(in crate::reader::codex) fn to_persisted(&self) -> PersistedUserTurnSlot {
         PersistedUserTurnSlot {
             blocks: self.blocks.clone(),
             preceding_message_id: self.preceding_message_id.clone(),
@@ -239,51 +234,54 @@ impl UserTurnSlot {
 }
 
 #[derive(Debug, Clone)]
-struct SpawnCallInfo {
-    call_id: String,
-    ts: String,
-    subagent_type: Option<String>,
-    description: Option<String>,
-    spawned_agent_id: Option<String>,
-    emitted: bool,
+pub(in crate::reader::codex) struct SpawnCallInfo {
+    pub(in crate::reader::codex) call_id: String,
+    pub(in crate::reader::codex) ts: String,
+    pub(in crate::reader::codex) subagent_type: Option<String>,
+    pub(in crate::reader::codex) description: Option<String>,
+    pub(in crate::reader::codex) spawned_agent_id: Option<String>,
+    pub(in crate::reader::codex) emitted: bool,
 }
 
 #[derive(Debug, Clone)]
-struct OpenTurn {
-    turn_id: String,
-    ts: String,
-    model: String,
-    project: Option<String>,
-    start_cumulative: CumulativeUsage,
-    tool_calls: Vec<ToolCall>,
-    seen_call_ids: BTreeSet<String>,
-    files_touched: BTreeSet<String>,
-    user_text: String,
-    assistant_text: String,
-    errored_call_ids: BTreeSet<String>,
-    content: Vec<ContentRecord>,
-    pending_tool_result_events: Vec<ToolResultEventRecord>,
-    pending_relationships: Vec<SessionRelationshipRecord>,
-    spawn_calls: HashMap<String, SpawnCallInfo>,
-    usage_observed: bool,
+pub(in crate::reader::codex) struct OpenTurn {
+    pub(in crate::reader::codex) turn_id: String,
+    pub(in crate::reader::codex) ts: String,
+    pub(in crate::reader::codex) model: String,
+    pub(in crate::reader::codex) project: Option<String>,
+    pub(in crate::reader::codex) start_cumulative: CumulativeUsage,
+    pub(in crate::reader::codex) tool_calls: Vec<ToolCall>,
+    pub(in crate::reader::codex) seen_call_ids: BTreeSet<String>,
+    pub(in crate::reader::codex) files_touched: BTreeSet<String>,
+    pub(in crate::reader::codex) user_text: String,
+    pub(in crate::reader::codex) assistant_text: String,
+    pub(in crate::reader::codex) errored_call_ids: BTreeSet<String>,
+    pub(in crate::reader::codex) content: Vec<ContentRecord>,
+    pub(in crate::reader::codex) pending_tool_result_events: Vec<ToolResultEventRecord>,
+    pub(in crate::reader::codex) pending_relationships: Vec<SessionRelationshipRecord>,
+    pub(in crate::reader::codex) spawn_calls: HashMap<String, SpawnCallInfo>,
+    pub(in crate::reader::codex) usage_observed: bool,
 }
 
-struct FinalizedTurn {
-    turn_id: String,
-    ts: String,
-    model: String,
-    project: Option<String>,
-    tool_calls: Vec<ToolCall>,
-    files_touched: Vec<String>,
-    user_text: String,
-    assistant_text: String,
-    errored_call_ids: BTreeSet<String>,
-    content: Vec<ContentRecord>,
-    usage: Usage,
-    fidelity: Fidelity,
+pub(in crate::reader::codex) struct FinalizedTurn {
+    pub(in crate::reader::codex) turn_id: String,
+    pub(in crate::reader::codex) ts: String,
+    pub(in crate::reader::codex) model: String,
+    pub(in crate::reader::codex) project: Option<String>,
+    pub(in crate::reader::codex) tool_calls: Vec<ToolCall>,
+    pub(in crate::reader::codex) files_touched: Vec<String>,
+    pub(in crate::reader::codex) user_text: String,
+    pub(in crate::reader::codex) assistant_text: String,
+    pub(in crate::reader::codex) errored_call_ids: BTreeSet<String>,
+    pub(in crate::reader::codex) content: Vec<ContentRecord>,
+    pub(in crate::reader::codex) usage: Usage,
+    pub(in crate::reader::codex) fidelity: Fidelity,
 }
 
-fn finalize_turn(open: OpenTurn, cumulative: &CumulativeUsage) -> FinalizedTurn {
+pub(in crate::reader::codex) fn finalize_turn(
+    open: OpenTurn,
+    cumulative: &CumulativeUsage,
+) -> FinalizedTurn {
     let usage = Usage {
         input: (cumulative.input - open.start_cumulative.input).max(0) as u64,
         output: (cumulative.output - open.start_cumulative.output).max(0) as u64,
@@ -331,992 +329,10 @@ fn build_codex_fidelity(usage_observed: bool) -> Fidelity {
 }
 
 // ---------------------------------------------------------------------------
-// Core parser loop
-// ---------------------------------------------------------------------------
-
-#[derive(Debug, Clone)]
-struct Pending<T> {
-    offset: u64,
-    record: T,
-}
-
-/// Owns the mutable WORKING state of the Codex parse pass. The driver
-/// (`parse_codex_buffer`) keeps the parallel `committed_*` shadow set and
-/// snapshots fields off this struct at `task_complete` commit boundaries, so
-/// a trailing partial/un-terminated line's mutations are discarded. The final
-/// result is assembled from the committed snapshots, never from this working
-/// state directly.
-struct CodexParseState {
-    session_id: String,
-    session_cwd: Option<String>,
-    turn_contexts: HashMap<String, CodexTurnContext>,
-    cumulative: CumulativeUsage,
-    open_turn: Option<OpenTurn>,
-    pending_user_text: String,
-    pending_content: Vec<ContentRecord>,
-    finalized: Vec<FinalizedTurn>,
-    user_turn_slot: UserTurnSlot,
-    user_turns: Vec<UserTurnRecord>,
-    root_session_emitted: bool,
-    seen_session_meta_keys: BTreeSet<String>,
-    next_event_index: u64,
-    tool_result_counters: HashMap<String, u64>,
-    last_completed_turn: Option<CodexLastCompletedTurn>,
-    pending_tool_result_events: Vec<Pending<ToolResultEventRecord>>,
-    pending_relationships: Vec<Pending<SessionRelationshipRecord>>,
-    pending_compactions: Vec<Pending<CompactionEvent>>,
-}
-
-impl CodexParseState {
-    /// Initialize working state from the resume option, mirroring the prior
-    /// inline `resume.map(...)` initializers verbatim.
-    fn new(resume: Option<&CodexResumeState>) -> Self {
-        Self {
-            session_id: resume.map(|r| r.session_id.clone()).unwrap_or_default(),
-            session_cwd: resume.and_then(|r| r.session_cwd.clone()),
-            turn_contexts: resume.map(|r| r.turn_contexts.clone()).unwrap_or_default(),
-            cumulative: resume.map(|r| r.cumulative.clone()).unwrap_or_default(),
-            open_turn: None,
-            pending_user_text: String::new(),
-            pending_content: Vec::new(),
-            finalized: Vec::new(),
-            user_turn_slot: resume
-                .and_then(|r| r.user_turn_slot.as_ref())
-                .map(UserTurnSlot::from_persisted)
-                .unwrap_or_default(),
-            user_turns: Vec::new(),
-            root_session_emitted: resume.map(|r| r.root_session_emitted).unwrap_or(false),
-            seen_session_meta_keys: resume
-                .map(|r| r.session_meta_relationship_keys.iter().cloned().collect())
-                .unwrap_or_default(),
-            next_event_index: resume.map(|r| r.next_event_index).unwrap_or(0),
-            tool_result_counters: resume
-                .map(|r| r.tool_result_counters.clone())
-                .unwrap_or_default(),
-            last_completed_turn: resume.and_then(|r| r.last_completed_turn.clone()),
-            pending_tool_result_events: Vec::new(),
-            pending_relationships: Vec::new(),
-            pending_compactions: Vec::new(),
-        }
-    }
-
-    fn handle_session_meta(&mut self, payload: &Value, rec_timestamp: &str, line_end_offset: u64) {
-        if let Some(id) = session_meta_payload_id(payload) {
-            self.session_id = id;
-        }
-        if let Some(cwd) = payload.get("cwd").and_then(|v| v.as_str()) {
-            self.session_cwd = Some(cwd.to_string());
-            if let Some(open) = self.open_turn.as_mut() {
-                if open.project.is_none() {
-                    open.project = Some(cwd.to_string());
-                }
-            }
-        }
-        if !self.session_id.is_empty() && !self.root_session_emitted {
-            self.root_session_emitted = true;
-            let ts = payload
-                .get("timestamp")
-                .and_then(|v| v.as_str())
-                .unwrap_or(rec_timestamp);
-            self.pending_relationships.push(Pending {
-                offset: line_end_offset,
-                record: build_root_relationship(&self.session_id, ts, payload),
-            });
-        }
-        if !self.session_id.is_empty() {
-            for row in build_session_meta_relationships(&self.session_id, payload, rec_timestamp) {
-                let key = codex_relationship_key(&row);
-                if self.seen_session_meta_keys.contains(&key) {
-                    continue;
-                }
-                self.seen_session_meta_keys.insert(key);
-                self.pending_relationships.push(Pending {
-                    offset: line_end_offset,
-                    record: row,
-                });
-            }
-        }
-    }
-
-    fn handle_turn_context(&mut self, payload: &Value) {
-        let ctx = CodexTurnContext {
-            turn_id: payload
-                .get("turn_id")
-                .and_then(|v| v.as_str())
-                .map(str::to_string),
-            cwd: payload
-                .get("cwd")
-                .and_then(|v| v.as_str())
-                .map(str::to_string),
-            model: payload
-                .get("model")
-                .and_then(|v| v.as_str())
-                .map(str::to_string),
-        };
-        if let Some(tid) = ctx.turn_id.clone() {
-            self.turn_contexts.insert(tid.clone(), ctx.clone());
-            if let Some(open) = self.open_turn.as_mut() {
-                if open.turn_id == tid {
-                    if open.model.is_empty() {
-                        if let Some(m) = ctx.model.as_deref() {
-                            open.model = m.to_string();
-                        }
-                    }
-                    if open.project.is_none() {
-                        if let Some(c) = ctx.cwd.as_deref() {
-                            open.project = Some(c.to_string());
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    fn handle_compacted(&mut self, rec_timestamp: &str, line_end_offset: u64) {
-        if !self.session_id.is_empty() {
-            self.pending_compactions.push(Pending {
-                offset: line_end_offset,
-                record: build_codex_compaction_event(
-                    &self.session_id,
-                    rec_timestamp,
-                    self.last_completed_turn.as_ref(),
-                ),
-            });
-        }
-    }
-
-    fn handle_event_msg(
-        &mut self,
-        payload: &Value,
-        rec_timestamp: &str,
-        capture_content: bool,
-        line_end_offset: u64,
-        committed: &mut CommittedSnapshot,
-    ) {
-        let pl_type = payload.get("type").and_then(|v| v.as_str()).unwrap_or("");
-        match pl_type {
-            "token_count" => {
-                if let Some(total) = payload.get("info").and_then(|i| {
-                    if i.is_null() {
-                        None
-                    } else {
-                        i.get("total_token_usage")
-                    }
-                }) {
-                    let input_total = total
-                        .get("input_tokens")
-                        .and_then(|v| v.as_i64())
-                        .unwrap_or(0);
-                    let cached = total
-                        .get("cached_input_tokens")
-                        .and_then(|v| v.as_i64())
-                        .unwrap_or(0);
-                    self.cumulative.input = input_total - cached;
-                    self.cumulative.cache_read = cached;
-                    self.cumulative.output = total
-                        .get("output_tokens")
-                        .and_then(|v| v.as_i64())
-                        .unwrap_or(0);
-                    self.cumulative.reasoning = total
-                        .get("reasoning_output_tokens")
-                        .and_then(|v| v.as_i64())
-                        .unwrap_or(0);
-                    if let Some(open) = self.open_turn.as_mut() {
-                        open.usage_observed = true;
-                    }
-                }
-            }
-            "task_started" => {
-                let ts = rec_timestamp;
-                let turn_id = match payload.get("turn_id").and_then(|v| v.as_str()) {
-                    Some(t) => t.to_string(),
-                    None => return,
-                };
-                if let Some(open) = self.open_turn.take() {
-                    self.finalized.push(finalize_turn(open, &self.cumulative));
-                }
-                if !self.user_turn_slot.blocks.is_empty() {
-                    self.user_turns.push(build_codex_user_turn_record(
-                        &self.user_turn_slot,
-                        &self.session_id,
-                        &turn_id,
-                        ts,
-                    ));
-                }
-                self.user_turn_slot = UserTurnSlot::default();
-                let ctx = self.turn_contexts.get(&turn_id).cloned();
-                let project = ctx
-                    .as_ref()
-                    .and_then(|c| c.cwd.clone())
-                    .or_else(|| self.session_cwd.clone());
-                let mut open = OpenTurn {
-                    turn_id: turn_id.clone(),
-                    ts: ts.to_string(),
-                    model: ctx
-                        .as_ref()
-                        .and_then(|c| c.model.clone())
-                        .unwrap_or_default(),
-                    project,
-                    start_cumulative: self.cumulative.clone(),
-                    tool_calls: vec![],
-                    seen_call_ids: BTreeSet::new(),
-                    files_touched: BTreeSet::new(),
-                    user_text: std::mem::take(&mut self.pending_user_text),
-                    assistant_text: String::new(),
-                    errored_call_ids: BTreeSet::new(),
-                    content: vec![],
-                    pending_tool_result_events: vec![],
-                    pending_relationships: vec![],
-                    spawn_calls: HashMap::new(),
-                    usage_observed: false,
-                };
-                if capture_content && !self.pending_content.is_empty() {
-                    for c in self.pending_content.iter_mut() {
-                        c.message_id = turn_id.clone();
-                    }
-                    open.content.append(&mut self.pending_content);
-                }
-                self.open_turn = Some(open);
-            }
-            "task_complete" => {
-                let payload_turn_id = payload
-                    .get("turn_id")
-                    .and_then(|v| v.as_str())
-                    .unwrap_or("");
-                let mut took: Option<OpenTurn> = None;
-                if let Some(open) = self.open_turn.as_ref() {
-                    if open.turn_id == payload_turn_id {
-                        took = self.open_turn.take();
-                    }
-                }
-                if let Some(mut open) = took {
-                    // Patch isError on tool-result blocks accumulated this turn
-                    for b in self.user_turn_slot.blocks.iter_mut() {
-                        if matches!(b.kind, UserTurnBlockKind::ToolResult) {
-                            if let Some(id) = &b.tool_use_id {
-                                if open.errored_call_ids.contains(id) {
-                                    b.is_error = Some(true);
-                                }
-                            }
-                        }
-                    }
-                    // Drain pending tool result events / relationships
-                    let mut events = std::mem::take(&mut open.pending_tool_result_events);
-                    for ev in events.iter_mut() {
-                        if open.errored_call_ids.contains(&ev.tool_use_id) {
-                            ev.status = ToolResultStatus::Errored;
-                            ev.is_error = Some(true);
-                        } else if matches!(ev.status, ToolResultStatus::Unknown) {
-                            ev.status = ToolResultStatus::Completed;
-                        }
-                    }
-                    for ev in events {
-                        self.pending_tool_result_events.push(Pending {
-                            offset: line_end_offset,
-                            record: ev,
-                        });
-                    }
-                    let rels = std::mem::take(&mut open.pending_relationships);
-                    for r in rels {
-                        self.pending_relationships.push(Pending {
-                            offset: line_end_offset,
-                            record: r,
-                        });
-                    }
-                    self.user_turn_slot.preceding_message_id = Some(open.turn_id.clone());
-                    let closed = finalize_turn(open, &self.cumulative);
-                    self.last_completed_turn = Some(CodexLastCompletedTurn {
-                        message_id: closed.turn_id.clone(),
-                        cache_read: closed.usage.cache_read,
-                    });
-                    self.finalized.push(closed);
-                    // Commit snapshot
-                    committed.snapshot(self, line_end_offset);
-                }
-            }
-            "patch_apply_end" => {
-                if let Some(open) = self.open_turn.as_mut() {
-                    let turn_id = payload.get("turn_id").and_then(|v| v.as_str());
-                    if turn_id != Some(open.turn_id.as_str()) {
-                        return;
-                    }
-                    let success = payload.get("success").and_then(|v| v.as_bool());
-                    if success == Some(false) {
-                        if let Some(call_id) = payload.get("call_id").and_then(|v| v.as_str()) {
-                            open.errored_call_ids.insert(call_id.to_string());
-                        }
-                        return;
-                    }
-                    if let Some(changes) = payload.get("changes").and_then(|v| v.as_object()) {
-                        for file in changes.keys() {
-                            open.files_touched.insert(file.clone());
-                        }
-                    }
-                }
-            }
-            "exec_command_end" => {
-                if let Some(open) = self.open_turn.as_mut() {
-                    let turn_id = payload.get("turn_id").and_then(|v| v.as_str());
-                    if turn_id != Some(open.turn_id.as_str()) {
-                        return;
-                    }
-                    let exit_code = payload.get("exit_code").and_then(|v| v.as_i64());
-                    if let (Some(code), Some(call_id)) =
-                        (exit_code, payload.get("call_id").and_then(|v| v.as_str()))
-                    {
-                        if code != 0 {
-                            open.errored_call_ids.insert(call_id.to_string());
-                        }
-                    }
-                }
-            }
-            other if is_subagent_terminal_notification(other) => {
-                let call_id = match payload.get("call_id").and_then(|v| v.as_str()) {
-                    Some(c) if !c.is_empty() => c.to_string(),
-                    _ => return,
-                };
-                let entry = self
-                    .tool_result_counters
-                    .entry(call_id.clone())
-                    .or_insert(0);
-                let call_index = *entry;
-                *entry += 1;
-                let status = subagent_notification_status(payload);
-                let mut ev = ToolResultEventRecord {
-                    v: 1,
-                    source: SourceKind::Codex,
-                    session_id: self.session_id.clone(),
-                    message_id: self.open_turn.as_ref().map(|o| o.turn_id.clone()),
-                    tool_use_id: call_id.clone(),
-                    call_index: Some(call_index),
-                    event_index: self.next_event_index,
-                    ts: if rec_timestamp.is_empty() {
-                        None
-                    } else {
-                        Some(rec_timestamp.to_string())
-                    },
-                    status,
-                    event_source: ToolResultEventSource::SubagentNotification,
-                    content_length: None,
-                    output_bytes: None,
-                    output_truncated: None,
-                    content_hash: None,
-                    is_error: matches!(status, ToolResultStatus::Errored).then_some(true),
-                    usage: None,
-                    usage_attribution: None,
-                    subagent_session_id: None,
-                    agent_id: None,
-                    replaced_tools: None,
-                    collapsed_calls: None,
-                };
-                self.next_event_index += 1;
-                let spawned_id =
-                    pick_string_field(payload, &["agent_id", "subagent_id", "session_id"]);
-                if let Some(sid) = spawned_id.as_ref() {
-                    ev.agent_id = Some(sid.clone());
-                    ev.subagent_session_id = Some(sid.clone());
-                    if let Some(open) = self.open_turn.as_mut() {
-                        if let Some(spawn) = open.spawn_calls.get_mut(&call_id) {
-                            if spawn.spawned_agent_id.is_none() {
-                                spawn.spawned_agent_id = Some(sid.clone());
-                                let info = spawn.clone();
-                                maybe_emit_spawn_relationship(
-                                    open,
-                                    &self.session_id,
-                                    &info,
-                                    rec_timestamp,
-                                );
-                            }
-                        }
-                    }
-                }
-                if let Some(open) = self.open_turn.as_mut() {
-                    open.pending_tool_result_events.push(ev);
-                } else {
-                    self.pending_tool_result_events.push(Pending {
-                        offset: line_end_offset,
-                        record: ev,
-                    });
-                }
-            }
-            _ => {}
-        }
-    }
-
-    fn handle_response_item(
-        &mut self,
-        payload: &Value,
-        rec_timestamp: &str,
-        capture_content: bool,
-        line_end_offset: u64,
-        counter: &HeuristicCounter,
-    ) {
-        let item_ts = rec_timestamp;
-        let pl_type = payload.get("type").and_then(|v| v.as_str()).unwrap_or("");
-        match pl_type {
-            "message" => {
-                let role = payload
-                    .get("role")
-                    .and_then(|v| v.as_str())
-                    .unwrap_or("")
-                    .to_string();
-                let text = collect_message_text(payload, &role);
-                if text.is_empty() {
-                    return;
-                }
-                if role == "user" {
-                    if let Some(open) = self.open_turn.as_mut() {
-                        open.user_text = append_text(&open.user_text, &text);
-                    } else {
-                        self.pending_user_text = append_text(&self.pending_user_text, &text);
-                    }
-                    self.user_turn_slot
-                        .blocks
-                        .push(UserTurnBlock::text(&text, counter));
-                    if self.user_turn_slot.ts.is_empty() && !item_ts.is_empty() {
-                        self.user_turn_slot.ts = item_ts.to_string();
-                    }
-                    if capture_content {
-                        let rec = ContentRecord {
-                            v: 1,
-                            source: SourceKind::Codex,
-                            session_id: self.session_id.clone(),
-                            message_id: self
-                                .open_turn
-                                .as_ref()
-                                .map(|o| o.turn_id.clone())
-                                .unwrap_or_default(),
-                            ts: item_ts.to_string(),
-                            role: ContentRole::User,
-                            kind: ContentKind::Text,
-                            text: Some(text.clone()),
-                            tool_use: None,
-                            tool_result: None,
-                        };
-                        push_content(&mut self.open_turn, &mut self.pending_content, rec);
-                    }
-                } else if role == "assistant" {
-                    if let Some(open) = self.open_turn.as_mut() {
-                        open.assistant_text = append_text(&open.assistant_text, &text);
-                        if capture_content {
-                            open.content.push(ContentRecord {
-                                v: 1,
-                                source: SourceKind::Codex,
-                                session_id: self.session_id.clone(),
-                                message_id: open.turn_id.clone(),
-                                ts: item_ts.to_string(),
-                                role: ContentRole::Assistant,
-                                kind: ContentKind::Text,
-                                text: Some(text),
-                                tool_use: None,
-                                tool_result: None,
-                            });
-                        }
-                    }
-                }
-            }
-            "reasoning" => {
-                if !capture_content {
-                    return;
-                }
-                let Some(open) = self.open_turn.as_mut() else {
-                    return;
-                };
-                let text = collect_reasoning_text(payload);
-                if !text.is_empty() {
-                    open.content.push(ContentRecord {
-                        v: 1,
-                        source: SourceKind::Codex,
-                        session_id: self.session_id.clone(),
-                        message_id: open.turn_id.clone(),
-                        ts: item_ts.to_string(),
-                        role: ContentRole::Assistant,
-                        kind: ContentKind::Thinking,
-                        text: Some(text),
-                        tool_use: None,
-                        tool_result: None,
-                    });
-                }
-            }
-            "function_call_output" | "custom_tool_call_output" => {
-                let call_id = match payload.get("call_id").and_then(|v| v.as_str()) {
-                    Some(c) => c.to_string(),
-                    None => return,
-                };
-                let output = payload.get("output").cloned().unwrap_or(Value::Null);
-                self.user_turn_slot.blocks.push(UserTurnBlock::tool_result(
-                    call_id.clone(),
-                    &output,
-                    None,
-                    counter,
-                ));
-                if self.user_turn_slot.ts.is_empty() && !item_ts.is_empty() {
-                    self.user_turn_slot.ts = item_ts.to_string();
-                }
-                let entry = self
-                    .tool_result_counters
-                    .entry(call_id.clone())
-                    .or_insert(0);
-                let call_index = *entry;
-                *entry += 1;
-                let initial_status = if self
-                    .open_turn
-                    .as_ref()
-                    .map(|o| o.errored_call_ids.contains(&call_id))
-                    .unwrap_or(false)
-                {
-                    ToolResultStatus::Errored
-                } else {
-                    ToolResultStatus::Unknown
-                };
-                let measured = measure_tool_output(&output);
-                let mut ev = ToolResultEventRecord {
-                    v: 1,
-                    source: SourceKind::Codex,
-                    session_id: self.session_id.clone(),
-                    message_id: self.open_turn.as_ref().map(|o| o.turn_id.clone()),
-                    tool_use_id: call_id.clone(),
-                    call_index: Some(call_index),
-                    event_index: self.next_event_index,
-                    ts: if item_ts.is_empty() {
-                        None
-                    } else {
-                        Some(item_ts.to_string())
-                    },
-                    status: initial_status,
-                    event_source: ToolResultEventSource::FunctionCallOutput,
-                    content_length: measured.length,
-                    output_bytes: measured.byte_length,
-                    // Codex doesn't carry an explicit truncation marker
-                    // distinct from its general output; leave None until
-                    // we have a concrete signal to flip on.
-                    output_truncated: None,
-                    content_hash: measured.hash,
-                    is_error: matches!(initial_status, ToolResultStatus::Errored).then_some(true),
-                    usage: None,
-                    usage_attribution: None,
-                    subagent_session_id: None,
-                    agent_id: None,
-                    replaced_tools: None,
-                    collapsed_calls: None,
-                };
-                self.next_event_index += 1;
-                if let Some(open) = self.open_turn.as_mut() {
-                    if let Some(spawn) = open.spawn_calls.get_mut(&call_id) {
-                        if let Some(sid) = extract_spawned_agent_id(&output) {
-                            spawn.spawned_agent_id = Some(sid.clone());
-                            ev.agent_id = Some(sid.clone());
-                            ev.subagent_session_id = Some(sid);
-                        }
-                        let info = spawn.clone();
-                        maybe_emit_spawn_relationship(open, &self.session_id, &info, item_ts);
-                    }
-                    open.pending_tool_result_events.push(ev);
-                } else {
-                    self.pending_tool_result_events.push(Pending {
-                        offset: line_end_offset,
-                        record: ev,
-                    });
-                }
-                if capture_content {
-                    let rec = ContentRecord {
-                        v: 1,
-                        source: SourceKind::Codex,
-                        session_id: self.session_id.clone(),
-                        message_id: self
-                            .open_turn
-                            .as_ref()
-                            .map(|o| o.turn_id.clone())
-                            .unwrap_or_default(),
-                        ts: item_ts.to_string(),
-                        role: ContentRole::ToolResult,
-                        kind: ContentKind::ToolResult,
-                        text: None,
-                        tool_use: None,
-                        tool_result: Some(ContentToolResult {
-                            tool_use_id: call_id,
-                            content: output,
-                            is_error: None,
-                        }),
-                    };
-                    push_content(&mut self.open_turn, &mut self.pending_content, rec);
-                }
-            }
-            "function_call" => {
-                let Some(open) = self.open_turn.as_mut() else {
-                    return;
-                };
-                let name = match payload.get("name").and_then(|v| v.as_str()) {
-                    Some(n) => n.to_string(),
-                    None => return,
-                };
-                let call_id = match payload.get("call_id").and_then(|v| v.as_str()) {
-                    Some(c) => c.to_string(),
-                    None => return,
-                };
-                if open.seen_call_ids.contains(&call_id) {
-                    return;
-                }
-                open.seen_call_ids.insert(call_id.clone());
-                let arg_str = payload.get("arguments").and_then(|v| v.as_str());
-                let parsed_args = arg_str.and_then(safe_parse_json_object);
-                let hash_input = parsed_args
-                    .clone()
-                    .map(Value::Object)
-                    .unwrap_or_else(|| Value::Object(Default::default()));
-                let target = pick_function_call_target(&name, parsed_args.as_ref());
-                let call = ToolCall {
-                    id: call_id.clone(),
-                    name: name.clone(),
-                    target,
-                    args_hash: args_hash(&hash_input),
-                    is_error: None,
-                    edit_pre_hash: None,
-                    edit_post_hash: None,
-                    skill_name: None,
-                    replaced_tools: None,
-                    collapsed_calls: None,
-                };
-                open.tool_calls.push(call);
-                if name == "spawn_agent" {
-                    let mut info = SpawnCallInfo {
-                        call_id: call_id.clone(),
-                        ts: item_ts.to_string(),
-                        subagent_type: None,
-                        description: None,
-                        spawned_agent_id: None,
-                        emitted: false,
-                    };
-                    if let Some(args) = parsed_args.as_ref() {
-                        let v = Value::Object(args.clone());
-                        info.subagent_type =
-                            pick_string_field(&v, &["subagent_type", "agent_type", "type"]);
-                        info.description =
-                            pick_string_field(&v, &["description", "task", "prompt"]);
-                        info.spawned_agent_id =
-                            pick_string_field(&v, &["agent_id", "subagent_id", "session_id"]);
-                    }
-                    open.spawn_calls.insert(call_id.clone(), info.clone());
-                    maybe_emit_spawn_relationship(open, &self.session_id, &info, item_ts);
-                }
-                if capture_content {
-                    let input = parsed_args
-                        .map(|m| m.into_iter().collect())
-                        .unwrap_or_default();
-                    open.content.push(ContentRecord {
-                        v: 1,
-                        source: SourceKind::Codex,
-                        session_id: self.session_id.clone(),
-                        message_id: open.turn_id.clone(),
-                        ts: item_ts.to_string(),
-                        role: ContentRole::Assistant,
-                        kind: ContentKind::ToolUse,
-                        text: None,
-                        tool_use: Some(ContentToolUse {
-                            id: call_id,
-                            name,
-                            input,
-                        }),
-                        tool_result: None,
-                    });
-                }
-            }
-            "custom_tool_call" => {
-                let Some(open) = self.open_turn.as_mut() else {
-                    return;
-                };
-                let name = match payload.get("name").and_then(|v| v.as_str()) {
-                    Some(n) => n.to_string(),
-                    None => return,
-                };
-                let call_id = match payload.get("call_id").and_then(|v| v.as_str()) {
-                    Some(c) => c.to_string(),
-                    None => return,
-                };
-                if open.seen_call_ids.contains(&call_id) {
-                    return;
-                }
-                open.seen_call_ids.insert(call_id.clone());
-                let input = payload
-                    .get("input")
-                    .and_then(|v| v.as_str())
-                    .unwrap_or("")
-                    .to_string();
-                let hash_input = serde_json::json!({ "input": input });
-                let target = pick_custom_tool_target(&name, &input);
-                let call = ToolCall {
-                    id: call_id.clone(),
-                    name: name.clone(),
-                    target,
-                    args_hash: args_hash(&hash_input),
-                    is_error: None,
-                    edit_pre_hash: None,
-                    edit_post_hash: None,
-                    skill_name: None,
-                    replaced_tools: None,
-                    collapsed_calls: None,
-                };
-                open.tool_calls.push(call);
-                if capture_content {
-                    let mut input_map = BTreeMap::new();
-                    input_map.insert("input".to_string(), Value::String(input));
-                    open.content.push(ContentRecord {
-                        v: 1,
-                        source: SourceKind::Codex,
-                        session_id: self.session_id.clone(),
-                        message_id: open.turn_id.clone(),
-                        ts: item_ts.to_string(),
-                        role: ContentRole::Assistant,
-                        kind: ContentKind::ToolUse,
-                        text: None,
-                        tool_use: Some(ContentToolUse {
-                            id: call_id,
-                            name,
-                            input: input_map,
-                        }),
-                        tool_result: None,
-                    });
-                }
-            }
-            _ => {}
-        }
-    }
-}
-
-/// The parallel `committed_*` shadow set. Snapshots advance only at
-/// `task_complete` boundaries; the final result is assembled from these
-/// fields so a trailing partial line's working-state mutations are dropped.
-struct CommittedSnapshot {
-    end_offset: u64,
-    cumulative: CumulativeUsage,
-    session_id: String,
-    session_cwd: Option<String>,
-    turn_contexts: HashMap<String, CodexTurnContext>,
-    finalized_count: usize,
-    user_turns_count: usize,
-    user_turn_slot: UserTurnSlot,
-    root_session_emitted: bool,
-    seen_session_meta_keys: BTreeSet<String>,
-    next_event_index: u64,
-    tool_result_counters: HashMap<String, u64>,
-    last_completed_turn: Option<CodexLastCompletedTurn>,
-}
-
-impl CommittedSnapshot {
-    fn initial(state: &CodexParseState, start_offset: u64) -> Self {
-        Self {
-            end_offset: start_offset,
-            cumulative: state.cumulative.clone(),
-            session_id: state.session_id.clone(),
-            session_cwd: state.session_cwd.clone(),
-            turn_contexts: state.turn_contexts.clone(),
-            finalized_count: 0,
-            user_turns_count: 0,
-            user_turn_slot: state.user_turn_slot.clone(),
-            root_session_emitted: state.root_session_emitted,
-            seen_session_meta_keys: state.seen_session_meta_keys.clone(),
-            next_event_index: state.next_event_index,
-            tool_result_counters: state.tool_result_counters.clone(),
-            last_completed_turn: state.last_completed_turn.clone(),
-        }
-    }
-
-    /// Advance the committed snapshot to the current working state at a
-    /// `task_complete` commit boundary. Mirrors the prior inline assignments
-    /// verbatim.
-    fn snapshot(&mut self, state: &CodexParseState, line_end_offset: u64) {
-        self.end_offset = line_end_offset;
-        self.cumulative = state.cumulative.clone();
-        self.session_id = state.session_id.clone();
-        self.session_cwd = state.session_cwd.clone();
-        self.turn_contexts = state.turn_contexts.clone();
-        self.finalized_count = state.finalized.len();
-        self.user_turns_count = state.user_turns.len();
-        self.user_turn_slot = state.user_turn_slot.clone();
-        self.root_session_emitted = state.root_session_emitted;
-        self.seen_session_meta_keys = state.seen_session_meta_keys.clone();
-        self.next_event_index = state.next_event_index;
-        self.tool_result_counters = state.tool_result_counters.clone();
-        self.last_completed_turn = state.last_completed_turn.clone();
-    }
-}
-
-fn parse_codex_buffer<R: BufRead>(
-    mut reader: R,
-    start_offset: u64,
-    options: &ParseCodexIncrementalOptions,
-    project_resolver: &ProjectResolver,
-) -> std::io::Result<ParseCodexIncrementalResult> {
-    let capture_content = matches!(options.content_mode, Some(ContentStoreMode::Full));
-    // Validated by `resolve_token_counter` at the public entry point.
-    let counter = HeuristicCounter;
-
-    let mut state = CodexParseState::new(options.resume.as_ref());
-    let mut committed = CommittedSnapshot::initial(&state, start_offset);
-
-    let mut line_buf: Vec<u8> = Vec::new();
-    let mut current_offset: u64 = start_offset;
-    loop {
-        line_buf.clear();
-        let n = reader.read_until(b'\n', &mut line_buf)?;
-        if n == 0 {
-            break;
-        }
-        // Drop trailing partial lines — the next incremental call resumes
-        // from the committed end offset, which only advances past `\n`.
-        if line_buf.last() != Some(&b'\n') {
-            break;
-        }
-        let line_end_offset = current_offset + n as u64;
-        current_offset = line_end_offset;
-        let text = std::str::from_utf8(&line_buf[..n - 1]).unwrap_or("").trim();
-        if text.is_empty() {
-            continue;
-        }
-        let parsed: Value = match serde_json::from_str(text) {
-            Ok(v) => v,
-            Err(_) => continue,
-        };
-        if !parsed.is_object() {
-            continue;
-        }
-        let rec_type = parsed.get("type").and_then(|v| v.as_str()).unwrap_or("");
-        let rec_timestamp = parsed
-            .get("timestamp")
-            .and_then(|v| v.as_str())
-            .unwrap_or("");
-        let payload = match parsed.get("payload") {
-            Some(p) if p.is_object() => p,
-            _ => continue,
-        };
-
-        match rec_type {
-            "session_meta" => state.handle_session_meta(payload, rec_timestamp, line_end_offset),
-            "turn_context" => state.handle_turn_context(payload),
-            "compacted" => state.handle_compacted(rec_timestamp, line_end_offset),
-            "event_msg" => state.handle_event_msg(
-                payload,
-                rec_timestamp,
-                capture_content,
-                line_end_offset,
-                &mut committed,
-            ),
-            "response_item" => state.handle_response_item(
-                payload,
-                rec_timestamp,
-                capture_content,
-                line_end_offset,
-                &counter,
-            ),
-            _ => continue,
-        }
-    }
-
-    // Emit only committed turns.
-    let committed_turns = &state.finalized[..committed.finalized_count];
-    let mut turns: Vec<TurnRecord> = Vec::with_capacity(committed_turns.len());
-    let mut content_out: Vec<ContentRecord> = Vec::new();
-    for (i, f) in committed_turns.iter().enumerate() {
-        let mut record = TurnRecord {
-            v: 1,
-            source: SourceKind::Codex,
-            session_id: committed.session_id.clone(),
-            session_path: options.session_path.clone(),
-            message_id: f.turn_id.clone(),
-            turn_index: i as u64,
-            ts: f.ts.clone(),
-            model: f.model.clone(),
-            project: None,
-            project_key: None,
-            usage: f.usage.clone(),
-            tool_calls: f.tool_calls.clone(),
-            files_touched: if f.files_touched.is_empty() {
-                None
-            } else {
-                Some(f.files_touched.clone())
-            },
-            subagent: None,
-            stop_reason: None,
-            activity: None,
-            retries: None,
-            has_edits: None,
-            fidelity: Some(f.fidelity.clone()),
-        };
-        if let Some(p) = f.project.as_ref() {
-            let resolved = project_resolver.resolve(p);
-            record.project = Some(resolved.project);
-            record.project_key = resolved.project_key;
-        }
-        let combined_text = join_nonempty(&[f.user_text.as_str(), f.assistant_text.as_str()], "\n");
-        let has_failed_tool = f
-            .tool_calls
-            .iter()
-            .any(|tc| f.errored_call_ids.contains(&tc.id));
-        let classified = classify_activity(ClassificationInput {
-            tool_calls: &f.tool_calls,
-            text: &combined_text,
-            has_failed_tool,
-            reasoning_tokens: f.usage.reasoning,
-        });
-        record.activity = Some(classified.activity);
-        record.retries = Some(classified.retries);
-        record.has_edits = Some(classified.has_edits);
-        turns.push(record);
-        if capture_content {
-            content_out.extend(f.content.clone());
-        }
-    }
-
-    let resume = CodexResumeState {
-        cumulative: committed.cumulative.clone(),
-        session_id: committed.session_id.clone(),
-        session_cwd: committed.session_cwd.clone(),
-        turn_contexts: committed.turn_contexts.clone(),
-        user_turn_slot: Some(committed.user_turn_slot.to_persisted()),
-        root_session_emitted: committed.root_session_emitted,
-        session_meta_relationship_keys: committed.seen_session_meta_keys.iter().cloned().collect(),
-        next_event_index: committed.next_event_index,
-        tool_result_counters: committed.tool_result_counters.clone(),
-        last_completed_turn: committed.last_completed_turn.clone(),
-    };
-
-    let user_turns_out = state.user_turns[..committed.user_turns_count].to_vec();
-    let mut events_out: Vec<CompactionEvent> = Vec::new();
-    for e in state.pending_compactions {
-        if e.offset <= committed.end_offset {
-            events_out.push(e.record);
-        }
-    }
-    let mut relationships_out: Vec<SessionRelationshipRecord> = Vec::new();
-    for r in state.pending_relationships {
-        if r.offset <= committed.end_offset {
-            relationships_out.push(r.record);
-        }
-    }
-    let mut tool_events_out: Vec<ToolResultEventRecord> = Vec::new();
-    for ev in state.pending_tool_result_events {
-        if ev.offset <= committed.end_offset {
-            tool_events_out.push(ev.record);
-        }
-    }
-
-    Ok(ParseCodexIncrementalResult {
-        turns,
-        content: content_out,
-        events: events_out,
-        user_turns: user_turns_out,
-        relationships: relationships_out,
-        tool_result_events: tool_events_out,
-        end_offset: committed.end_offset,
-        resume,
-    })
-}
-
-// ---------------------------------------------------------------------------
 // Helpers
 // ---------------------------------------------------------------------------
 
-fn session_meta_payload_id(payload: &Value) -> Option<String> {
+pub(in crate::reader::codex) fn session_meta_payload_id(payload: &Value) -> Option<String> {
     let id = payload.get("id")?.as_str()?;
     if id.is_empty() {
         None
@@ -1325,7 +341,7 @@ fn session_meta_payload_id(payload: &Value) -> Option<String> {
     }
 }
 
-fn collect_message_text(payload: &Value, role: &str) -> String {
+pub(in crate::reader::codex) fn collect_message_text(payload: &Value, role: &str) -> String {
     let Some(content) = payload.get("content").and_then(|v| v.as_array()) else {
         return String::new();
     };
@@ -1368,7 +384,7 @@ fn is_codex_boilerplate(text: &str) -> bool {
     false
 }
 
-fn collect_reasoning_text(payload: &Value) -> String {
+pub(in crate::reader::codex) fn collect_reasoning_text(payload: &Value) -> String {
     let mut parts: Vec<String> = Vec::new();
     if let Some(arr) = payload.get("summary").and_then(|v| v.as_array()) {
         for s in arr {
@@ -1391,7 +407,7 @@ fn collect_reasoning_text(payload: &Value) -> String {
     parts.join("\n")
 }
 
-fn append_text(existing: &str, next: &str) -> String {
+pub(in crate::reader::codex) fn append_text(existing: &str, next: &str) -> String {
     if existing.is_empty() {
         next.to_string()
     } else {
@@ -1399,7 +415,9 @@ fn append_text(existing: &str, next: &str) -> String {
     }
 }
 
-fn safe_parse_json_object(s: &str) -> Option<serde_json::Map<String, Value>> {
+pub(in crate::reader::codex) fn safe_parse_json_object(
+    s: &str,
+) -> Option<serde_json::Map<String, Value>> {
     if s.is_empty() {
         return None;
     }
@@ -1410,7 +428,7 @@ fn safe_parse_json_object(s: &str) -> Option<serde_json::Map<String, Value>> {
     }
 }
 
-fn pick_function_call_target(
+pub(in crate::reader::codex) fn pick_function_call_target(
     name: &str,
     args: Option<&serde_json::Map<String, Value>>,
 ) -> Option<String> {
@@ -1429,7 +447,7 @@ fn pick_function_call_target(
     }
 }
 
-fn pick_custom_tool_target(name: &str, input: &str) -> Option<String> {
+pub(in crate::reader::codex) fn pick_custom_tool_target(name: &str, input: &str) -> Option<String> {
     if name != "apply_patch" {
         return None;
     }
@@ -1455,7 +473,7 @@ fn pick_custom_tool_target(name: &str, input: &str) -> Option<String> {
     None
 }
 
-fn pick_string_field(value: &Value, keys: &[&str]) -> Option<String> {
+pub(in crate::reader::codex) fn pick_string_field(value: &Value, keys: &[&str]) -> Option<String> {
     let obj = value.as_object()?;
     for k in keys {
         if let Some(s) = obj.get(*k).and_then(|v| v.as_str()) {
@@ -1467,7 +485,7 @@ fn pick_string_field(value: &Value, keys: &[&str]) -> Option<String> {
     None
 }
 
-fn extract_spawned_agent_id(output: &Value) -> Option<String> {
+pub(in crate::reader::codex) fn extract_spawned_agent_id(output: &Value) -> Option<String> {
     let obj_owner;
     let obj = match output {
         Value::Object(_) => output,
@@ -1484,17 +502,17 @@ fn extract_spawned_agent_id(output: &Value) -> Option<String> {
 }
 
 #[derive(Default)]
-struct Measured {
-    length: Option<u64>,
-    hash: Option<String>,
+pub(in crate::reader::codex) struct Measured {
+    pub(in crate::reader::codex) length: Option<u64>,
+    pub(in crate::reader::codex) hash: Option<String>,
     /// Raw UTF-8 byte length of the materialized payload. Same value as
     /// `length` for Codex (the legacy `content_length` already counted
     /// bytes here, not chars), but tracked separately so the
     /// `ToolResultEventRecord` shape stays consistent across sources.
-    byte_length: Option<u64>,
+    pub(in crate::reader::codex) byte_length: Option<u64>,
 }
 
-fn measure_tool_output(output: &Value) -> Measured {
+pub(in crate::reader::codex) fn measure_tool_output(output: &Value) -> Measured {
     match output {
         Value::Null => Measured::default(),
         Value::String(s) => Measured {
@@ -1513,7 +531,7 @@ fn measure_tool_output(output: &Value) -> Measured {
     }
 }
 
-fn is_subagent_terminal_notification(t: &str) -> bool {
+pub(in crate::reader::codex) fn is_subagent_terminal_notification(t: &str) -> bool {
     if !t.starts_with("subagent_") {
         return false;
     }
@@ -1523,7 +541,7 @@ fn is_subagent_terminal_notification(t: &str) -> bool {
         || t.ends_with("_terminated")
 }
 
-fn subagent_notification_status(payload: &Value) -> ToolResultStatus {
+pub(in crate::reader::codex) fn subagent_notification_status(payload: &Value) -> ToolResultStatus {
     if let Some(b) = payload.get("success").and_then(|v| v.as_bool()) {
         return if b {
             ToolResultStatus::Completed
@@ -1546,7 +564,11 @@ fn subagent_notification_status(payload: &Value) -> ToolResultStatus {
     ToolResultStatus::Completed
 }
 
-fn build_root_relationship(session_id: &str, ts: &str, meta: &Value) -> SessionRelationshipRecord {
+pub(in crate::reader::codex) fn build_root_relationship(
+    session_id: &str,
+    ts: &str,
+    meta: &Value,
+) -> SessionRelationshipRecord {
     let mut row = SessionRelationshipRecord {
         v: 1,
         source: RelationshipSourceKind::Codex,
@@ -1569,7 +591,7 @@ fn build_root_relationship(session_id: &str, ts: &str, meta: &Value) -> SessionR
     row
 }
 
-fn build_session_meta_relationships(
+pub(in crate::reader::codex) fn build_session_meta_relationships(
     session_id: &str,
     meta: &Value,
     fallback_ts: &str,
@@ -1639,7 +661,7 @@ fn apply_codex_session_meta_provenance(row: &mut SessionRelationshipRecord, meta
     }
 }
 
-fn codex_relationship_key(row: &SessionRelationshipRecord) -> String {
+pub(in crate::reader::codex) fn codex_relationship_key(row: &SessionRelationshipRecord) -> String {
     let source = match row.source {
         RelationshipSourceKind::Codex => "codex",
         RelationshipSourceKind::ClaudeCode => "claude-code",
@@ -1668,7 +690,7 @@ fn codex_relationship_key(row: &SessionRelationshipRecord) -> String {
     )
 }
 
-fn maybe_emit_spawn_relationship(
+pub(in crate::reader::codex) fn maybe_emit_spawn_relationship(
     open_turn: &mut OpenTurn,
     session_id: &str,
     info: &SpawnCallInfo,
@@ -1709,7 +731,7 @@ fn maybe_emit_spawn_relationship(
     }
 }
 
-fn push_content(
+pub(in crate::reader::codex) fn push_content(
     open_turn: &mut Option<OpenTurn>,
     pending: &mut Vec<ContentRecord>,
     record: ContentRecord,
@@ -1721,7 +743,7 @@ fn push_content(
     }
 }
 
-fn build_codex_user_turn_record(
+pub(in crate::reader::codex) fn build_codex_user_turn_record(
     slot: &UserTurnSlot,
     session_id: &str,
     following_message_id: &str,
@@ -1746,7 +768,7 @@ fn build_codex_user_turn_record(
     }
 }
 
-fn build_codex_compaction_event(
+pub(in crate::reader::codex) fn build_codex_compaction_event(
     session_id: &str,
     ts: &str,
     preceding: Option<&CodexLastCompletedTurn>,
@@ -1796,6 +818,14 @@ fn clone_resume(r: Option<&CodexResumeState>) -> CodexResumeState {
 // AgentWorkforce/burn#430 and the module's own doc comment for the
 // scope discussion.
 pub mod span_tree;
+
+// Incremental parse engine: the `CodexParseState` streaming state machine,
+// its `CommittedSnapshot` shadow, and the `parse_codex_buffer` driver the
+// public `parse_codex_session*` entry points wrap. Split out of this file; the
+// helpers/types above feed it per line.
+mod incremental;
+
+use self::incremental::parse_codex_buffer;
 
 #[cfg(test)]
 mod tests;
